- Fix wrong variable of `Ion.Beam` used for `IonCannon` - Fix wrong variable used for `FistrormWall` link updating - Fix crash with `TemporalClass::GetWarpPerStep` hook - Fix more compiler warnings - Fix `DriverKilled` unit was driving around randomly because missing code - Fix internal `SpeedMult` causing unit to stuck inside warfactory - Fix `GiftBox` resetting it duration when the techno already dead - Fix `LineTrail` not properly recreated after `TypeConversion` - Fix `Spawner` properties not changed after `TypeConversion` - Fix `DisguiseWhenStill` properties not changed after `TypeConversion` - Fix bunch of potentially dangerious `Uninitialized` variables - Fix crash related to `RadSite` that caused by leftover code - Fix crash related to `BulletClass` with `Cluster` - using own `RegisteredJammers` map for better memory handling - add prototype `Sequence` reding code for the future - add `Bounty.ReceiveSound` - add `Rules->General->Everyone.CanDrive` global toggle - add `AI.CloneAs` - `Aircraft` `SmokeAnim` now can be disable completely without dummy anim - no longer allow invalid aircraft type to be used as `SpyPlane` and `ParadropPlane` - make AI unit not become passive when it is failed to find path when hunt mission - replace some `UniequePtrB` to `Handle<T,Deleter>` for better handling - remove hook  that causing building failed to render on campaign - rename some variables - sync to latest phobos - Known Crash or bugs:    - Late game crash caused by some unknown in progress    - Load game sometime crash with EIP `0x00000001` when loading game inside game    - `DetonateOnAllMapObjects` can crash the game when internal game array unexpectedly changed    - maybe more ,..

+++ src/Utilities/GeneralUtils.cpp
@@ -259,6 +259,6 @@ int GeneralUtils::GetLSAnimHeightFactor(AnimTypeClass* pType, CellClass* pCell, 
 	const auto Height = int(((double)ImageHeignt) / 2);
 	const auto LSFactor = (Height - 0.5) * LightningStorm::CloudHeightFactor();
 
-	return int(checklevel ? LSFactor + Unsorted::LevelHeight * pCell->Level : LSFactor);
+	return int(checklevel ? LSFactor + (double)Unsorted::LevelHeight * (double)pCell->Level : LSFactor);
 }
 #pragma endregion
